Evil Armor (#1951)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds Zizo and Matthios armor.

![image](https://github.com/user-attachments/assets/3d405c2c-9de7-4b43-b027-2c8089d8be0b)

![image](https://github.com/user-attachments/assets/fe2a1442-1fc2-442f-af2c-f2ddeda9d2a6)

## Why It's Good For The Game

Ook and Gale wanted me to add this, not acquirable currently ingame but
still available for adminspawn and future PRs.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->