New Assassin Disguises (#1961)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds four new Assassin disguises.

Thief - Same as Adventurer.

Ranger - Same as Adventurer.

Servant - Does not possess a Manor Key.

Assassin - A full-white outfit complete with sunglasses (resprited from
Inquisitor sunglasses, thank you cannibal hunter for your color
wizardry.) Makes the character bald, because bald people are generally
evil. Hunt Showdown Night Seer my beloved.


![BALD2](https://github.com/user-attachments/assets/c2048148-f3de-474c-b51a-eae7e2ebad9a)


## Why It's Good For The Game

As it stands, Hunter and Noble are the most useful disguises - having a
bow and a way into the keep, respectively. Many players who take Hunted
also tend to end up in the Keep. You tend to see a good amount of failed
assassinations from the rafters.

This won't exactly tackle that (I feel it's good to mention it) but it
will give more variety to hopefully avoid these scenarios where people
see one of these roles and instantly know. They're disguises for a
reason. We need more of them.

Having the option to play the obvious neighborhood serial killer is also
nice.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->