[FOLLOWUP] Access is handled by a lock datum (#1722)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

# Follow-up of #1441

## About The Pull Request


Access moved to a lock datum with obj/procs for interacting with the
lock, Lockpicking unelementized and moved to the lock.

Locks come in two kinds:
keyless lock which expect the object to handle locking and unlocking
like a deadbolt door.
keylock which uses attackby signals to handle keys and lockpick
attempts.

This means that all key unlocking is left click now.

Lockid has been converted to a list so that locks can be more flexible.

This fixes the issue with locks always having a novice lock.

MASTER OF NERVES, PEDDLERs, GOLDFACE, PURITY can all be lockpicked.
(PEDDLER has to handle item insertion via rightclick now so we don't
have to mess around with intents or combat mode to enable this)

Access helpers can be applied to closets, PEDDLERs and doors. Instead of
just doors.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Objects mostly don't need to reinvent the wheel when it comes to
handling access.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: Arkatos1 <43862960+Arkatos1@users.noreply.github.com>