Fixes Aasimar swordmaster spawning without their armor, remove dwarf from being aple to play sword master(no armor sprite for them) (#1952)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Makes so Aasimar can equip the grezen armor and remove dwarf from the
possibility to roll that role
Dwarves can't equip the armor at the moment and even if they could, the
sprite would not fit because of the pixel difference

![AASIMAR_armor](https://github.com/user-attachments/assets/3b79eaba-fa50-4764-8adb-820ddd1f09a9)



## Why It's Good For The Game

Fix

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+grungussuss@users.noreply.github.com>