Warden Balancing (#1816)
Removes dodge expert, as Warden wears heavy armor. Reduced coinpurse
size to middle tier, as they are a noble as reward, not as lifestyle.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
--> Fixes some power creep issues on Warden:
- Removes Dodge Expert as they have heavy armor
- Removes some of their starting money to discourage round-start smithy
line sitting.
- Removes medium armor training as they have a heavy armor focus
- UPDATED: -1 strength, +1 endurance and athletics.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. --> Changes Warden from doom guy into a marginally more
veteran soldier than the rest of his foresters.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->