Zizoid Knight (#1980)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Gives Drow Knight the full Zizo set and a Zizo sword (Ported from Azure
Peak) added in this PR.

![image](https://github.com/user-attachments/assets/f7dac24c-b487-4357-b5b9-2ff9869e6db6)

## Why It's Good For The Game

Frees up the armor to be used for bladesinger, now people wont look at
Drow Knight and go "Oh thats just a drow bladesinger" instead thinking
"Oh fuck that guy is EVIL"

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->