Minor Update (#2022)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

![image](https://github.com/user-attachments/assets/71a7c39c-80fc-4686-8dcb-578693e55c18)

![image](https://github.com/user-attachments/assets/0a6a26de-b88a-4036-b0f3-4596a0bcc215)

![image](https://github.com/user-attachments/assets/112a68f9-7fc9-4593-acf2-501c79aca0d2)

![image](https://github.com/user-attachments/assets/52068933-ca41-4f4c-905a-2cecf42e8785)

Map update provides new AI additions now that AI is fixed. Forest map on
rosewood hosts a maze dungeon that has fae only monsters in it. Maze has
a few little goodies at the end for the players to find.

Map also adds in vendors to the stewards office so that they can put
keys up for sale!

## Why It's Good For The Game

Adds more diversity to rosewoods otherz map and overall making the world
feel more full.

Also might solve the issue of the key situation with no one being able
to buy houses from steward due to steward being too busy.
## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->