Gives Assassin (disguise) a Tossbelt + Zigs + Tunic (#2005)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Gives the Assassin (disguise, not the job) the following:

A Zig Box
Flint
Noble Coat (Enhances the drip.)
STEEL Tossbelt


## Why It's Good For The Game

Tossbelt requested by Ook.

They get a steel tossblade belt since apparently they're the same as
iron I guess. Congratulations, he's a ninja or some shit now.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->